refactor: clean up code and break up large factory file
- Remove useless comments ('Creates the factory function...')
- Break workflow-step-tools.factory.ts (512 lines → 75 lines)
- Extract builders to separate files in builders/ directory:
  - create-record-step.builder.ts (94 lines)
  - update-record-step.builder.ts (101 lines)
  - find-records-step.builder.ts (149 lines)
  - delete-record-step.builder.ts (87 lines)
  - step-builder.utils.ts (52 lines)

All files now under 300 lines
