+++ packages/twenty-server/src/engine/core-modules/record-crud/tool-factory/direct-record-tools.factory.ts
@@ -22,7 +22,6 @@ export type DirectRecordToolsDeps = {
   findRecordsService: FindRecordsService;
 };
 
-// Creates the factory function with injected dependencies
 export const createDirectRecordToolsFactory = (deps: DirectRecordToolsDeps) => {
   return (
     {
